==> Notebooks/05_Modeling_with_Doc2Vec.ipynb <==
# Modeling

from sklearn.linear_model import LogisticRegression 
from sklearn.naive_bayes import GaussianNB, BernoulliNB
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC

from sklearn.model_selection import train_test_split, KFold, cross_val_score
from sklearn.metrics import confusion_matrix, classification_report, accuracy_score

from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV

from sklearn.manifold import TSNE
import matplotlib.pyplot as plt

import pickle

%matplotlib inline

### Import X and y

with open('../Data/X_doc2vec.pkl', 'rb') as f:
    X = pickle.load(f)

with open('../Data/y_doc2vec.pkl', 'rb') as f:
    y = pickle.load(f)

### Import Doc2Vec Models

with open('../Models/model_dbow.pkl', 'rb') as f:
    model_dbow = pickle.load(f)

with open('../Models/model_dm_mean.pkl', 'rb') as f:
    model_dm_mean = pickle.load(f)

with open('../Models/model_dm_concat.pkl', 'rb') as f:
    model_dm_concat = pickle.load(f)

models = [(model_dbow, 'model_dbow'), (model_dm_mean, 'model_dm_mean'), (model_dm_concat, 'model_dm_concat')]

With the vectors I got from the Doc2Vec models, I will fit a number of differt classifiers and evaluate the results.

### Bernoulli Naive_Bayes

def bern_nb_model_func(X, y):
    X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.25,random_state=42)
    bern_nb = BernoulliNB()
    bern_nb.fit(X_train, y_train)
    print("cross-validated train scores:", cross_val_score(bern_nb, X_train, y_train, cv=3))
    print("cross-validated test scores:", cross_val_score(bern_nb, X_test, y_test, cv=3))
    print('---')

for model in models:
    print(model)
    bern_nb_model_func(X[model[1]], y)

### Gaussian Naive-Bayes

def gb_model_func(X, y):
    X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.25,random_state=42)
    gb = GaussianNB()
    gb.fit(X_train, y_train)
    print("cross-validated train scores:", cross_val_score(gb, X_train, y_train, cv=3))
    print("cross-validated test scores:", cross_val_score(gb, X_test, y_test, cv=3))
    print('---')

for model in models:
    print(model)
    gb_model_func(X[model[1]], y)

### Logistic Regression

def lr_model_func(X,y):
    lr = LogisticRegression()
    pipe = Pipeline([
        ('lr', lr)
    ])
    
    params = {
        'lr__penalty': ['l1'],
        'lr__C': [0.5]
    }
    
    X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.25,random_state=42)
    gs_d2v_lr = GridSearchCV(pipe, param_grid=params, cv=2)
    gs_d2v_lr.fit(X_train, y_train)
    print('train score:', gs_d2v_lr.score(X_train, y_train))
    print('test score:', gs_d2v_lr.score(X_test, y_test))
    print('best score:', gs_d2v_lr.best_score_)
    print('best params:', gs_d2v_lr.best_params_)
    print('---')

for model in models:
    print(model)
    lr_model_func(X[model[1]], y)

### Random Forest

def rf_model_func(X,y):
    rf = RandomForestClassifier()
    pipe = Pipeline([
        ('rf', rf)
    ])
    
    params = {
        'rf__n_estimators': [750],
        'rf__max_features': ['log2', 'sqrt']
    }
    
    X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.25,random_state=42)
    gs_d2v_rf = GridSearchCV(pipe, param_grid=params, cv=2)
    gs_d2v_rf.fit(X_train, y_train)
    print('train score:', gs_d2v_rf.score(X_train, y_train))
    print('test score:', gs_d2v_rf.score(X_test, y_test))
    print('best score:', gs_d2v_rf.best_score_)
    print('best params:', gs_d2v_rf.best_params_)
    print('---')

%%time

for model in models:
    print(model)
    rf_model_func(X[model[1]], y)

### Gradient Boost

def grad_model_func(X,y):
    grad = GradientBoostingClassifier()
    pipe = Pipeline([
        ('grad', grad)
    ])
    
    params = {
        'grad__n_estimators': [1000],
        'grad__max_features': ['log2']
    }
    
    X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.25,random_state=42)
    gs_d2v_grad = GridSearchCV(pipe, param_grid=params, cv=2)
    gs_d2v_grad.fit(X_train, y_train)
    print('train score:', gs_d2v_grad.score(X_train, y_train))
    print('test score:', gs_d2v_grad.score(X_test, y_test))
    print('best score:', gs_d2v_grad.best_score_)
    print('best params:', gs_d2v_grad.best_params_)
#     print('classification report': 
    print('---')

%%time

for model in models:
    print(model)
    grad_model_func(X[model[1]], y)

---

When pairing the Doc2Vec models with classifiers, the two pairings that produced the best scores were Doc2Vec DBOW with logistic regression and Doc2Vec DBOW with gradient boost.  I decided to use the gradient boost classifier in my production model.  I made this decision because gradient boosting is an ensemble technique which means it has many different predictors trying to predict the same target variable, and as a boosting method, it adds one classifier at a time so that the next classifier is trained to improve the already trained ensemble.  Logistic regression is a linear classifier and is not equipped to make the types of decision that gradient boosting does.
