==> Notebooks/06_tweet_scraper.ipynb <==
### Import Modules

from twython import Twython  
import json
import pandas as pd
import pickle

from twython import Twython  
import json
import pandas as pd
import pickle

#pass security information to variables
consumer_key = "XVFBqWVkhDlRBFz2f7CbvIiXI"
consumer_secret = "S4BO3WQpFHRX3q0vOTGZvre3TIPyxtt6w3WTPAexqNtii5fsIw"
access_key = "1034835017371770880-b4eFRilbAVFpPwA5tTXRp2JBhr6rWK"
access_secret = "VWAg8gzTHPXJxfY9GwGkcI2iuplAtw5NxhhnQPcbNwKRr"

# Instantiate an object
apple_tweets = Twython(consumer_key, consumer_secret)

# Create our query
query = {'q': '#FirstMan',  
        'result_type': 'recent',
        'lang': 'en',
        'count': 100,
        'tweet_mode':'extended'
        }

dict_ = {'user': [], 'date': [], 'text': []}  
npages = 150
for i in range(npages):
    # Search tweets
    print('Processing {0} out {1} pages'.format(i,npages))
    results = apple_tweets.search(**query)
    
    
    for status in results['statuses']:  
        dict_['user'].append(status['user']['screen_name'])
        dict_['date'].append(status['created_at'])
        if status.get('retweeted_status'):
            dict_['text'].append(status['retweeted_status']['full_text'])
        else:
            dict_['text'].append(status['full_text'])
            
    query['max_id'] = results['statuses'][-1]['id_str']

**Structure data in a pandas DataFrame for easier manipulation**

tweets_df = pd.DataFrame(dict_)  

**Convert date column into datetime format**

tweets_df['date'] = pd.to_datetime(tweets_df['date'])

**Drop duplicates**

tweets_df.drop_duplicates('text', inplace=True)

**Reset Index**

tweets_df.reset_index(inplace=True)

tweets_df.drop(columns='index', inplace=True)

tweets_df.shape

### Export tweets_df

with open('../Data/tweets_df', 'wb+') as f:
    pickle.dump(tweets_df, f)
