==> Notebooks/07_Application_of_Model.ipynb <==
# Application

This notebook is a demonstration of how to apply the sentiment analysis model I have built.  I have imported 5732 tweets from the past week, all containing the hashtag #FirstMan (in reference to a movie that was recently released).  The tweets will be processed as follows:

- They will be concatenated to the original movie review dataframe on the 'review' column
- They will be preprocessed by the review_cleaner function 
- Doc2Vec will train its models on the entire new corpus (movie reviews + tweets) in order to build a new vector space
- A gradient boost classifier will then train on these new vectors and will output a value of 'positive' or 'negative' when assessing the new tweets

### Import Modules

import pandas as pd
import numpy as np
from pprint import pprint

import re
import pickle

import requests
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
from sklearn.metrics import classification_report
from bs4 import BeautifulSoup
from nltk.tokenize import WordPunctTokenizer

from gensim.models.doc2vec import Doc2Vec, TaggedDocument

from collections import OrderedDict
import multiprocessing

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
from textblob import TextBlob

import warnings
warnings.filterwarnings('ignore')
from sklearn import utils

np.random.seed(42)

### Import tweets_df

with open('../Data/tweets_df', 'rb') as f:
    tweets_df = pickle.load(f)

tweets_df.head()

# DataFrame Concatenation

### Read Data into Pandas Table

**Perform all operations that were already done to original dataset.**

df = pd.read_csv('../Data/imdb_master.csv', encoding = "ISO-8859-1")
df.drop(columns='Unnamed: 0', axis=1, inplace=True)

df_train = df[df['type'] == 'train']
df_train.reset_index(inplace=True)
df_train.drop(columns=['index'], inplace=True)

df_train = df_train[df_train['label'] != 'unsup']

df_train.drop_duplicates('review', inplace=True)

df_train.shape

### Concatenate tweets_df to df_train

tweets_df = tweets_df.rename(columns = {'text':'review'})

df_train.shape

tweets_df.shape

df_train = pd.merge(df_train, tweets_df, how='outer', on='review')

df_train.head()

### Drop Rows

df_train.drop(columns=['type', 'file', 'user', 'date', 'user', 'date'], inplace=True)

df_train.head()

# Text Preprocessing

url = 'https://raw.githubusercontent.com/stanfordnlp/CoreNLP/master/data/edu/stanford/nlp/patterns/surface/stopwords.txt'
res = requests.get(url)
soup = BeautifulSoup(res.content, 'lxml')

standford_nlp = soup.text.split('\n')
stop_word_list = standford_nlp + list(ENGLISH_STOP_WORDS) + ['film', 'films', 'movie', 'movies']

tok = WordPunctTokenizer()

negations_dict = {"isn't":"is not", "aren't":"are not", "wasn't":"was not", "weren't":"were not",
                "haven't":"have not","hasn't":"has not","hadn't":"had not","won't":"will not",
                "wouldn't":"would not", "don't":"do not", "doesn't":"does not","didn't":"did not",
                "can't":"can not","couldn't":"could not","shouldn't":"should not","mightn't":"might not",
                "mustn't":"must not"}
neg_pattern = re.compile(r'\b(' + '|'.join(negations_dict.keys()) + r')\b')
emoji_pattern = re.compile("["
        u"\U0001F600-\U0001F64F"  # emoticons
        u"\U0001F300-\U0001F5FF"  # symbols & pictographs
        u"\U0001F680-\U0001F6FF"  # transport & map symbols
        u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           "]+", flags=re.UNICODE)

def review_cleaner(text):
    soup = BeautifulSoup(text, 'lxml')
    souped = soup.get_text()
    stripped  = re.sub(r'@[A-Za-z0-9_]+', '', souped)
    stripped = re.sub(r'https?://[^ ]+', '', stripped)
    stripped = re.sub(r'www.[^ ]+', '', stripped)
    lower_case = text.lower()
    neg_dict = neg_pattern.sub(lambda x: negations_dict[x.group()], lower_case)
    letters_only = re.sub('[^a-zA-Z]', " ", neg_dict)
    no_emoji = re.sub(emoji_pattern, '', letters_only)
    # During the letters_only process two lines above, it has created unnecessay white spaces,
    # I will tokenize and join together to remove unneccessary white spaces
    words = [x for x  in tok.tokenize(letters_only) if len(x) > 1]
    return (" ".join(words)).strip()

### Map Function

df_train['review'] = df_train['review'].map(review_cleaner)

df_train.head()

# Doc2Vec

tagged_documents = []
for indx, doc in enumerate(df_train["review"].values):
    tagged_documents.append(TaggedDocument([x for x in doc.split()], [indx]))

cores = multiprocessing.cpu_count()
vec_size = 150

model_dbow = Doc2Vec(dm=0, dbow_words=1, vector_size=vec_size, negative=5, hs=0, min_count=2, sample=0, 
             workers=cores)

model_dm_mean = Doc2Vec(dm=1, dm_mean=1, vector_size=vec_size, window=10, negative=5, hs=0, min_count=2, sample=0, 
                workers=cores, alpha=0.05, comment='alpha=0.05')

model_dm_concat = Doc2Vec(dm=1, dm_concat=1, vector_size=vec_size, window=5, negative=5, hs=0, min_count=2, sample=0, 
                  workers=cores)

models = [(model_dbow, 'model_dbow'), (model_dm_mean, 'model_dm_mean'), (model_dm_concat, 'model_dm_concat')]

for model in models:
    model[0].build_vocab(tagged_documents)
    print("%s vocabulary scanned & state initialized" % model[0])
    
models_by_name = OrderedDict((str(model[1]), model[0]) for model in models)

for model in models:
    for epoch in range(30):
        print('Epoch: {0}'.format(epoch), 'Model: %s' % (model[0]))
        model[0].train(utils.shuffle(tagged_documents), total_examples=len(tagged_documents), epochs=1)
        model[0].alpha -= 0.002
        model[0].min_alpha = model[0].alpha

### Define X and y

X = {}

for model in models:
    X[model[1]] = np.zeros((df_train.shape[0], vec_size))
    for i in range(df_train.shape[0]):
        X[model[1]][i] = model[0].docvecs[i]  

y = df_train['label'].values

# Modeling

from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier

from sklearn.model_selection import train_test_split, KFold, cross_val_score
from sklearn.metrics import confusion_matrix, classification_report, accuracy_score

from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV

import matplotlib.pyplot as plt
%matplotlib inline

### Gradient Boost

grad_model = {}

def grad_model_func(X,y):
    grad = GradientBoostingClassifier()
    pipe = Pipeline([
        ('grad', grad)
    ])
    
    params = {
        'grad__n_estimators': [1000],
        'grad__max_features': ['log2']
    }
    
    X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.25,random_state=42)
    gs_d2v_grad = GridSearchCV(pipe, param_grid=params, cv=2)
    gs_d2v_grad.fit(X_train, y_train)
    print('train score:', gs_d2v_grad.score(X_train, y_train))
    print('test score:', gs_d2v_grad.score(X_test, y_test))
    print('best score:', gs_d2v_grad.best_score_)
    print('best params:', gs_d2v_grad.best_params_)
    print('---')
    return gs_d2v_grad

for model in models:
    print(model)
    m = grad_model_func(X[model[1]][:24904,:], y[:24904])
    grad_model[model[1]+'_trained'] = m

Create new data frame with all new rows that have import tweets

df_new = df_train.iloc[24904:].reset_index(drop=True)
df_new.head()

Create a new column 'predicted' that will contain a prediction of positive or negative for each tweet.

df_train['predicted'] = grad_model['model_dbow_trained'].predict(X[models[0][1]])

Create a count of all predicted positive and negative tweets analyzed by my model.

pos_count = 0
neg_count = 0

for score in df_train.loc[24904:, 'predicted']:
    if score ==  'pos':
        pos_count += 1
    if score == 'neg':
        neg_count += 1
        
print("Positive count = {}".format(pos_count))
print("Negative count = {}".format(neg_count))

# TextBlob and VADER Sentiment

### TextBlob

Apply TextBlob.sentiment to all columns with tweets and designate output to new column (TextBlob.sentiment is scored on a scale of -1 to 1.)

df_new['textblob'] = df_new['review'].apply(lambda x: TextBlob(x).polarity)

df_new.head()

Create a count of all positive, neutral and negative tweets analyzed by TextBlob

pos_count = 0
neg_count = 0
neutral_count = 0

for score in df_new['textblob']:
    if score > 0:
        pos_count += 1


for score in df_new['textblob']:
    if score < 0:
        neg_count += 1


for score in df_new['textblob']:
    if score == 0:
        neutral_count += 1

print("Positive count = {}".format(pos_count))
print("Negative count = {}".format(neg_count))
print("Neutral count = {}".format(neutral_count))

### Vader Sentiment

Apply SentimentIntensityAnalyzer.polarity_scores to all columns with tweets and designate the ouput to a new column (Vader sentiment is scored on a scale of -4 to 4.)

analyzer = SentimentIntensityAnalyzer()

vs = analyzer.polarity_scores

df_new['vader_sentiment'] = df_new['review'].apply(lambda x: vs(x))

pos_count = 0
neg_count = 0
neutral_count = 0

for dictionary in df_new.loc[:, 'vader_sentiment']:
    if dictionary['compound'] > 0:
        pos_count += 1
    elif dictionary['compound'] < 0:
        neg_count += 1
    else:
        neutral_count += 1 

print("Positive count = {}".format(pos_count))
print("Negative count = {}".format(neg_count))
print("Neutral count = {}".format(neutral_count))

### Sentiment Count df

i = ['my_model', 'textblob', 'vader_sentiment']
s = {'positive': [5162, 3715, 3321], 'negative': [570, 717, 1007], 'neutral': ['XXX', 1300, 1404]}

pd.DataFrame(data=s, index=i)
